Gradle tests: check analyzer results in BaseIntegrationSpec
Verify analyzer results with expected results for all analyzer output
files in BaseIntegrationSpec.
